+++ tests/integrations/test_bottle.py
@@ -33,13 +33,12 @@ class TestBottle(IntegrationTest):
         self.assertEqual(event['context'], 'GET /beans')
         self.assertEqual(event['exceptions'][0]['errorClass'], 'Exception')
         self.assertEqual(event['exceptions'][0]['message'], 'oh no!')
-        self.assertEqual(event['metaData']['environment']['PATH_INFO'],
-                         '/beans')
         runtime_versions = event['device']['runtimeVersions']
         self.assertEqual(runtime_versions['bottle'], '0.12.18')
+        assert 'environment' not in event['metaData']
 
-    def test_disable_environment(self):
-        bugsnag.configure(send_environment=False)
+    def test_enable_environment(self):
+        bugsnag.configure(send_environment=True)
 
         @route('/beans')
         def index():
@@ -53,7 +52,7 @@ class TestBottle(IntegrationTest):
         self.assertEqual(1, len(self.server.received))
         payload = self.server.received[0]['json_body']
         metadata = payload['events'][0]['metaData']
-        assert 'environment' not in metadata
+        self.assertEqual(metadata['environment']['PATH_INFO'], '/beans')
 
     def test_template_error(self):
         @route('/berries/<variety>')
@@ -74,7 +73,6 @@ class TestBottle(IntegrationTest):
         self.assertEqual(event['exceptions'][0]['errorClass'], 'NameError')
         self.assertEqual(event['exceptions'][0]['message'],
                          "name 'type2' is not defined")
-        self.assertEqual(event['metaData']['environment']['PATH_INFO'],
-                         '/berries/red')
+        assert 'environment' not in event['metaData']
         runtime_versions = event['device']['runtimeVersions']
         self.assertEqual(runtime_versions['bottle'], bottle.__version__)